requirements.txt
beautifulsoup4==4.11.1
Pillow==9.2.0
PySide6==6.3.1
python_magic==0.4.27
